Move jsonrpc error codes to their own file
