+++ tests/connections.py
@@ -13,7 +13,6 @@ import multiprocessing
 from random import seed
 import infra.ccf
 import infra.proc
-import infra.client
 import json
 import contextlib
 import resource
