+++ tests/e2e_batched.py
@@ -6,7 +6,6 @@ import time
 
 import infra.ccf
 import infra.proc
-import infra.client
 import infra.notification
 import infra.net
 import suite.test_requirements as reqs
