+++ tests/e2e_logging_pbft.py
@@ -9,7 +9,6 @@ import shutil
 from random import seed
 import infra.ccf
 import infra.proc
-import infra.client
 import infra.notification
 import infra.net
 import e2e_args
